+++ src/utils/generateToken.go
@@ -9,7 +9,7 @@ import (
 
 func GenerateToken(id uint) (string, error) {
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
-		"sub": id,
+		"sub": int64(id),
 		"exp": time.Now().Add(time.Hour * 24).Unix(),
 	})
